requirements.txt
numpy
cv2
pillow
pytesseract
imutils
skimage